fix: resolve 71 compilation errors from split imports
- integration_handlers.rs: make all fns pub, add missing imports (Utc, types, INTEGRATION_JOIN_SQL, to_response)
- stream/mod.rs: remove duplicate mod handlers declaration
- stream/handlers.rs: make all fns pub, add missing imports (interval, Event, Sse, Stream, Duration, types)
- sso/config/mod.rs: remove stray derive attr, make VerifyDomainRequest field pub
- sso/config/domain.rs: add imports (ErrorCode, OidcDiscovery, TeamQuery, VerifyDomainRequest), add Serialize derive + pub fields to VerifyDomainResponse
- sso/scim/mod.rs: make scim_row_to_customer/scim_row_to_attrs pub
- sso/scim/users.rs: add import for scim_row_to_* functions
- background.rs: fix retention_days type i32→i64, add Datelike import

cargo check: 0 errors

+++ api/src/background.rs
@@ -4,6 +4,7 @@
 //! Uses distributed locks (via Redis) to prevent duplicate work
 //! in multi-instance deployments.
 
+use chrono::Datelike;
 use sqlx::PgPool;
 
 use crate::config;
@@ -20,7 +21,7 @@ pub fn spawn_background_jobs(
     job_queue: Option<jobs::job_queue::JobQueue>,
     email_provider: EmailProvider,
     fcm_client: Option<FcmClient>,
-    retention_days: i32,
+    retention_days: i64,
 ) {
     // Job queue worker (processes email + notification jobs from Redis)
     if let Some(ref queue) = job_queue {
